[MCUX-87600] Fix Coverity and MISRA issues in PBAP, MAP, L2CAP shell, and byteorder
- bt_pal_pbap_pce.c: Add overflow check before computing appl_param_len to
  prevent uint16_t overflow; add validation to prevent underflow when
  tag.id is 0; add bounds check before memcpy to prevent buffer overrun;
  fix event_result conversion using bt_pbap_convert_result() in callbacks
- bt_pal_pbap_pse.c: Remove unnecessary cast; reject phonebook names that
  exceed buffer size instead of silent truncation; initialize result
  variable only in valid switch cases; fix event_result conversion in
  disconnected callback
- bt_pal_map_mce.c: Add overflow check when updating pkt_len; add explicit
  cast to prevent implicit conversion warning; initialize result variable
  only in valid switch cases in bt_map_convert_result()
- bt_pal_map_mse.c: Add validation to prevent underflow when tag_bytes->id
  is 0; add explicit cast for tag.id assignment
- shell_l2cap.c: Simplify data rate calculation to avoid potential overflow
  (replace (x * 8) / 1024 with x / 128); add explicit casts to prevent
  uint16_t multiplication overflow in total_byte_count calculations
- fsl_shell.c: Add parentheses around *indicator dereference before addition
  to ensure correct order of operations in buffer overflow check
- byteorder.h: Add outer cast to ensure sys_get_le16() result is properly
  typed as uint16_t

These changes eliminate potential integer overflow, buffer overrun, and
implicit conversion defects identified by Coverity and MISRA analysis.

+++ include/sys/byteorder.h
@@ -540,7 +540,7 @@ static inline uint64_t sys_get_be64(const uint8_t src[8])
  */
 static inline uint16_t sys_get_le16(const uint8_t src[2])
 {
-	return ((uint16_t)src[1] << 8) | src[0];
+	return (uint16_t)(((uint16_t)src[1] << 8) | src[0]);
 }
 
 /**
